+++ gazelle/python/configure.go
@@ -70,6 +70,7 @@ func (py *Configurer) KnownDirectives() []string {
 		pythonconfig.LabelConvention,
 		pythonconfig.LabelNormalization,
 		pythonconfig.GeneratePyiDeps,
+		pythonconfig.GeneratePyiSrcs,
 		pythonconfig.ExperimentalAllowRelativeImports,
 		pythonconfig.GenerateProto,
 		pythonconfig.PythonResolveSiblingImports,
@@ -242,6 +243,12 @@ func (py *Configurer) Configure(c *config.Config, rel string, f *rule.File) {
 				log.Fatal(err)
 			}
 			config.SetGeneratePyiDeps(v)
+		case pythonconfig.GeneratePyiSrcs:
+			v, err := strconv.ParseBool(strings.TrimSpace(d.Value))
+			if err != nil {
+				log.Fatal(err)
+			}
+			config.SetGeneratePyiSrcs(v)
 		case pythonconfig.GenerateProto:
 			v, err := strconv.ParseBool(strings.TrimSpace(d.Value))
 			if err != nil {