+++ gazelle/python/kinds.go
@@ -51,6 +51,7 @@ var pyKinds = map[string]rule.KindInfo{
 		ResolveAttrs: map[string]bool{
 			"deps":     true,
 			"pyi_deps": true,
+			"pyi_srcs": true,
 		},
 	},
 	pyLibraryKind: {
@@ -68,6 +69,7 @@ var pyKinds = map[string]rule.KindInfo{
 		ResolveAttrs: map[string]bool{
 			"deps":     true,
 			"pyi_deps": true,
+			"pyi_srcs": true,
 		},
 	},
 	pyProtoLibraryKind: {
@@ -91,6 +93,7 @@ var pyKinds = map[string]rule.KindInfo{
 		ResolveAttrs: map[string]bool{
 			"deps":     true,
 			"pyi_deps": true,
+			"pyi_srcs": true,
 		},
 	},
 }